+++ nope/lib/views/directive/toolbar.php
@@ -1,0 +1,4 @@
+<div class="nope-toolbar">
+  <nope-model ng-model="selection.image" href="#/media?mimetype=image/" model="Media" icon="image" button="true" label="n!"></nope-model>
+  <nope-model ng-model="selection.page" href="#/content/page" icon="file-text-o" button="true" label="n!"></nope-model>
+</div>