+++ nope/lib/views/directive/model.php
@@ -55,7 +55,7 @@ $templateList = sprintf($template, 'dnd-nodrag', 'item');
 <div class="nope-model">
   <div ng-if="!button">
     <ul dnd-list="ngModel" class="list-group list-group-contents is-multiple" ng-show="ngModel && preview" ng-if="multiple">
-      <li class="list-group-item media" ng-repeat="item in ngModel track by $index" dnd-draggable="item" dnd-moved="ngModel.splice($index,1)">
+      <li class="list-group-item media" ng-repeat="item in ngModel" dnd-draggable="item" dnd-moved="ngModel.splice($index,1);">
         <i class="fa fa-bars handle"></i>
         <?php echo $templateList; ?>
         <div dnd-nodrag class="btn-group btn-group-xs toolbar">