+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Fixed block button on empty states.
 - Fixed `Query::__to` method called statically without `static` key (#39).
 - Fixed loader not resizing accordingly on window resize (#38).
+- Fixed broken `Gallery` widget on related `Media` query.
+- Fixed broken _drag'n'drop_ removing `track by $index` on `ng-repeat` (#41).
 
 ## v0.11.0 (2016-02-16)